+++ app/src/main/kotlin/com/pitchedapps/frost/web/FrostWebViewCore.kt
@@ -1,6 +1,7 @@
 package com.pitchedapps.frost.web
 
 import android.animation.ValueAnimator
+import android.annotation.SuppressLint
 import android.content.Context
 import android.support.v4.view.NestedScrollingChild
 import android.support.v4.view.NestedScrollingChildHelper
@@ -94,6 +95,7 @@ class FrostWebViewCore @JvmOverloads constructor(
      *
      * https://github.com/takahirom/webview-in-coordinatorlayout/blob/master/app/src/main/java/com/github/takahirom/webview_in_coodinator_layout/NestedWebView.java
      */
+    @SuppressLint("ClickableViewAccessibility")
     override fun onTouchEvent(ev: MotionEvent): Boolean {
         val event = MotionEvent.obtain(ev)
         val action = event.action